+++ src/deckgl-layers/sharedstreets-layer/sharedstreets-layer.js
@@ -26,14 +26,21 @@ import Pbf from "pbf";
 import * as geobuf from "geobuf";
 import dataProcessor from 'processors';
 import { processGeojson } from 'processors/data-processor';
+import { memoize } from 'utils/utils';
  
 export default class SharedstreetsLayer extends CompositeLayer {
   initializeState() {
     this.setState({
-      isTiledSampleDataLoaded: false
+      isTiledSampleDataLoaded: false,
+      oldLayerDataMaps: new Map()
     })
   }
 
+  /**
+   * return data in tile z-x-y
+   * @param {object} tile
+   * @returns {object} {allData} in KeplerGl allData format
+   */
   getTileData({ x, y, z }) {
     const fetchConfig = {
       method: "GET",
@@ -61,18 +68,36 @@ export default class SharedstreetsLayer extends CompositeLayer {
           isTiledSampleDataLoaded: true
         });
       }
-      const allData = getAllData(geoJson);
-      return {
-        geoJson,
-        allData
-      };
+      // convert geojson to a format Kepler expects.
+      const allData = processGeojson(geoJson).rows;
+      return allData
     });
   }
   
+  /**
+   * returns all deckGl layers rendering data in the viewport. 
+   * @param {*} subLayerProps 
+   * @param {*} tile current tile for this sub-layers 
+   */
   renderSubLayers(subLayerProps, tile) {
     const {layers, objectHovered, mapState, interactionConfig, layerVersion} = this.props;
+    const {oldLayerDataMaps} = this.state;
+    // layers are kepler layers rendering a subset of data. We can render other layers in the 
+    // viewport by giving different ids and data. 
     return layers && layers.map((layer, idx) => {
-      const data = this.getLayerData(layer, tile, layerVersion)
+      const layerDataId = `${layer.id}-${tile.z}-${tile.x}-${tile.y}`;
+      let oldLayerData;
+      if (oldLayerDataMaps.has(layerDataId)) {
+        oldLayerData = oldLayerDataMaps.get(layerDataId);
+        if (oldLayerData.data.length === 0) {
+          oldLayerData = undefined;
+        }
+      }
+
+      const data = this.getLayerData(layer, tile, layerVersion, oldLayerData);
+      oldLayerDataMaps.set(layerDataId, data);
+      this.setState({ oldLayerDataMaps });
+
       return layer.renderLayer({
         id: `${subLayerProps.id}-${layer.id}`,
         data,
@@ -95,44 +120,37 @@ export default class SharedstreetsLayer extends CompositeLayer {
     ];
   }
 
-  recomputeLayerData(keplerLayer, tile) {
+  /**
+   * Returns the updated deckGl layer data props for the given tile and kepler layer. 
+   * @param {*} keplerLayer 
+   * @param {*} tile 
+   */
+  recomputeLayerData(keplerLayer, tile, oldLayerData) {
     if (tile.isLoaded) {
-      return keplerLayer.formatLayerData([], tile._data.allData, Array.from(tile._data.allData.keys()), keplerLayer);
+      return keplerLayer.formatLayerData([], tile._data, Array.from(tile._data.keys()), oldLayerData, {sameData: true});
     }
     tile.data.then(data => {
       this.setLayerNeedsUpdate();
     })
-    return keplerLayer.formatLayerData([], [], [], keplerLayer);
+    return keplerLayer.formatLayerData([], [], [], oldLayerData);
   }
 
-  getLayerData(keplerLayer, tile, layerVersion) {
+  /**
+   * Returns the deckGl layer data prop for the given keplerLayer, layerVersion and tile.
+   * @param {*} keplerLayer 
+   * @param {*} tile 
+   * @param {*} layerVersion 
+   */
+  getLayerData(keplerLayer, tile, layerVersion, oldLayerData) {
     if (!tile.keplerCache) {
       tile.keplerCache = {};
     }
     if (!tile.keplerCache[keplerLayer.id]) {
       tile.keplerCache[keplerLayer.id] = memoize(this.recomputeLayerData.bind(this));
     }
-    const result = tile.keplerCache[keplerLayer.id](keplerLayer, tile, layerVersion, tile.isLoaded);
+    // recompute layer data iff any of the following arguments (keplerlayer, tile, layerVersion,
+    // tile.isLoaded) changes.
+    const result = tile.keplerCache[keplerLayer.id](keplerLayer, tile, oldLayerData, layerVersion, tile.isLoaded);
     return result;
-  }
-}
-
-function getAllData(rawData) {
-  const allData = processGeojson(rawData).rows;
-  return allData;
-}
-
-function memoize(recompute) {
-  let lastArgs = [];
-  let lastRes = null;
-  return (...args) => {
-    for (let i = 0; i < args.length; i++) {
-      if (args[i] !== lastArgs[i]) {
-        lastRes = recompute(...args);
-        lastArgs = args;
-        return lastRes;
-      }
-    }
-    return lastRes;
   }
 }
